backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
librosa
numpy
python-dotenv
google-genai
openai-whisper
torch
torchaudio
soundfile
pydub
pydantic
